refactor(console): extract FilterChip + CreatedFilterChip to shared module
Lift the chip primitive + the Created-at time-bucket chip out of
AgentsList into apps/console/src/components/FilterChip.tsx so every
list page can drop them in by import. CreatedFilterChip is fully
self-contained — owns its preset choice + custom-range bounds, just
surfaces the resolved {after, before} epoch-ms tuple via onChange so
the parent can ship it straight to a server query param.

Also drop Dashboard's full-screen BrandLoader. Stats already fall
back to "–" while loading; recent-sessions panel gets a 4-row
skeleton matching the live table. With keepPreviousData on TQ, the
page only flashes anything on truly cold cache; even then the
chrome (header / steps / stats grid) renders immediately instead
of disappearing behind a centered spinner.
