+++ apps/console/src/pages/AgentsList.tsx
@@ -1,19 +1,13 @@
 import { useEffect, useMemo, useState } from "react";
 import { useNavigate } from "react-router";
-import { CheckIcon, ChevronDownIcon, XIcon } from "lucide-react";
 
 import { useApi } from "../lib/api";
 import { useInfiniteApiQuery } from "../lib/useApiQuery";
 import { DataTable, type ColumnDef } from "../components/DataTable";
 import { FacetedFilter } from "../components/FacetedFilter";
+import { FilterChip, CreatedFilterChip } from "../components/FilterChip";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import { cn } from "@/lib/utils";
+import { PopoverContent } from "@/components/ui/popover";
 import type { ModelCard } from "@open-managed-agents/api-types";
 import type { AgentRecord as Agent } from "../types/agent";
 import { AgentFormDialog } from "./agents/AgentFormDialog";
@@ -38,12 +32,6 @@ type Runtime = {
   >;
 };
 
-// ── Filter primitives ────────────────────────────────────────────────
-// Inline (only used by this page right now). Extract to its own module
-// the second time a list page needs the same chip pattern. Per the
-// project's no-future-proofing rule, we don't bake a generic
-// "FilterRegistry" until that second caller materializes.
-
 type StatusValue = "any" | "active" | "archived";
 
 const STATUS_OPTIONS: { value: StatusValue; label: string }[] = [
@@ -51,139 +39,6 @@ const STATUS_OPTIONS: { value: StatusValue; label: string }[] = [
   { value: "active", label: "Active" },
   { value: "archived", label: "Archived" },
 ];
-
-/** Preset Created-at buckets. `after`/`before` returned in epoch ms;
- *  `null` means "no bound on this side". Driven by the chip's static
- *  options below. Generated lazily because `Date.now()` rolls forward
- *  while the page is open — recomputed every time the chip opens so
- *  "Today" picks up date changes after midnight. */
-type CreatedPreset =
-  | "any"
-  | "today"
-  | "last-hour"
-  | "last-day"
-  | "last-7d"
-  | "last-30d"
-  | "custom";
-
-const CREATED_PRESET_LABELS: Record<CreatedPreset, string> = {
-  any: "All time",
-  today: "Today",
-  "last-hour": "Last hour",
-  "last-day": "Last day",
-  "last-7d": "Last 7 days",
-  "last-30d": "Last 30 days",
-  custom: "Custom range",
-};
-
-function computePresetRange(
-  preset: Exclude<CreatedPreset, "any" | "custom">,
-): { after?: number; before?: number } {
-  const now = Date.now();
-  switch (preset) {
-    case "today": {
-      const start = new Date();
-      start.setHours(0, 0, 0, 0);
-      return { after: start.getTime() };
-    }
-    case "last-hour":
-      return { after: now - 60 * 60 * 1000 };
-    case "last-day":
-      return { after: now - 24 * 60 * 60 * 1000 };
-    case "last-7d":
-      return { after: now - 7 * 24 * 60 * 60 * 1000 };
-    case "last-30d":
-      return { after: now - 30 * 24 * 60 * 60 * 1000 };
-  }
-}
-
-/** Compact chip-style filter trigger. Decoration follows selection,
- *  same principle as the sidebar nav: nothing visible at rest, only
- *  the actually-selected chip gets the brand pill outline + bg. Idle
- *  chips are bare `label ▾` text so an empty filter row visually
- *  weighs nothing; the toolbar reads as "nothing's filtered" without
- *  having to scan every chip for a value.
- *
- *  Built on shadcn Popover (Radix + Floating UI under the hood) so
- *  PopoverContent can host anything — cmdk Command for the
- *  search-and-pick FacetedFilter (Status), a custom preset list with
- *  date inputs (Created), etc. Earlier iteration used DropdownMenu but
- *  cmdk Command nested inside Radix Menu fights for focus/keyboard
- *  control. Popover with the same Floating UI primitive gives identical
- *  auto-flip + collision behavior with no menu semantics in the way. */
-function FilterChip({
-  label,
-  active,
-  display,
-  onClear,
-  children,
-}: {
-  label: string;
-  active: boolean;
-  display?: string;
-  onClear?: () => void;
-  children: React.ReactNode;
-}) {
-  return (
-    <Popover>
-      <div
-        className={cn(
-          "inline-flex items-center gap-1 h-8 text-sm shrink-0 transition-colors",
-          active
-            ? "rounded-full border border-brand text-brand bg-brand-subtle"
-            : "text-fg-muted hover:text-fg",
-        )}
-      >
-        <PopoverTrigger asChild>
-          <button
-            type="button"
-            className={cn(
-              "inline-flex items-center gap-1 h-full outline-none",
-              active ? "pl-3 pr-2" : "px-2",
-            )}
-          >
-            <span className="font-medium">{label}</span>
-            {display && (
-              <>
-                <span className="text-fg-subtle">:</span>
-                <span>{display}</span>
-              </>
-            )}
-            {!active && <ChevronDownIcon className="size-3.5 opacity-60" />}
-          </button>
-        </PopoverTrigger>
-        {active && onClear && (
-          <button
-            type="button"
-            onClick={(e) => {
-              e.stopPropagation();
-              onClear();
-            }}
-            className="inline-flex items-center justify-center size-5 mr-1.5 rounded-full hover:bg-brand/10"
-            aria-label={`Clear ${label} filter`}
-          >
-            <XIcon className="size-3" />
-          </button>
-        )}
-      </div>
-      {children}
-    </Popover>
-  );
-}
-
-/** Format epoch ms → YYYY-MM-DD for the native `<input type="date">`
- *  value attribute. Date pickers store local-tz dates as strings; we
- *  convert back to ms via `new Date(str).getTime()` (also local). The
- *  precision mismatch (day vs ms) is fine because the chip presets
- *  themselves resolve to coarse boundaries (today 00:00, etc). */
-function msToDateInput(ms: number | undefined): string {
-  if (ms === undefined) return "";
-  const d = new Date(ms);
-  const yyyy = d.getFullYear();
-  const mm = String(d.getMonth() + 1).padStart(2, "0");
-  const dd = String(d.getDate()).padStart(2, "0");
-  return `${yyyy}-${mm}-${dd}`;
-}
 
 export function AgentsList() {
   const { api } = useApi();
@@ -201,37 +56,23 @@ export function AgentsList() {
   // → useInfiniteApiQuery resets to page 1 on params change → the list
   // reflects exactly what the server returned (no client-side faking).
   const [status, setStatus] = useState<StatusValue>("active");
-  const [createdPreset, setCreatedPreset] = useState<CreatedPreset>("any");
-  const [customAfter, setCustomAfter] = useState<number | undefined>();
-  const [customBefore, setCustomBefore] = useState<number | undefined>();
+  const [created, setCreated] = useState<{ after?: number; before?: number }>({});
   const [search, setSearch] = useState("");
-
-  // Derive the actual (after, before) bounds from the preset choice.
-  // `custom` reads from the two date-picker states; the other presets
-  // recompute from Date.now() so the bucket walks forward as time
-  // passes (Today rolls over at midnight, Last 7 days slides daily).
-  const { createdAfter, createdBefore } = useMemo(() => {
-    if (createdPreset === "any") return { createdAfter: undefined, createdBefore: undefined };
-    if (createdPreset === "custom")
-      return { createdAfter: customAfter, createdBefore: customBefore };
-    const r = computePresetRange(createdPreset);
-    return { createdAfter: r.after, createdBefore: r.before };
-  }, [createdPreset, customAfter, customBefore]);
 
   // Main agents table — paginated. Any change to these params resets to
   // page 0 automatically (paramsKey is part of the query key).
   const agentsParams = useMemo(
     () => ({
       status,
-      ...(createdAfter !== undefined
-        ? { created_after: new Date(createdAfter).toISOString() }
+      ...(created.after !== undefined
+        ? { created_after: new Date(created.after).toISOString() }
         : {}),
-      ...(createdBefore !== undefined
-        ? { created_before: new Date(createdBefore).toISOString() }
+      ...(created.before !== undefined
+        ? { created_before: new Date(created.before).toISOString() }
         : {}),
       ...(search ? { q: search } : {}),
     }),
-    [status, createdAfter, createdBefore, search],
+    [status, created.after, created.before, search],
   );
   const {
     items: agents,
@@ -350,8 +191,6 @@ export function AgentsList() {
   // only renders when the chip is in non-default state.
   const statusDisplay =
     status === "any" ? undefined : STATUS_OPTIONS.find((o) => o.value === status)?.label;
-  const createdDisplay =
-    createdPreset === "any" ? undefined : CREATED_PRESET_LABELS[createdPreset];
 
   const filters = (
     <>
@@ -380,80 +219,7 @@ export function AgentsList() {
         </PopoverContent>
       </FilterChip>
 
-      <FilterChip
-        label="Created"
-        active={createdPreset !== "any"}
-        display={createdDisplay}
-        onClear={() => {
-          setCreatedPreset("any");
-          setCustomAfter(undefined);
-          setCustomBefore(undefined);
-        }}
-      >
-        {/* Created uses a hand-rolled preset list + date inputs
-            rather than FacetedFilter because (a) the option list is
-            a fixed 7-item bucket palette where search adds zero
-            value, and (b) the Custom-range pair (From / To date
-            inputs) needs interactive children that cmdk Command's
-            list semantics don't accommodate cleanly. */}
-        <PopoverContent
-          align="start"
-          sideOffset={4}
-          collisionPadding={8}
-          className="w-60 p-1"
-        >
-          {(Object.keys(CREATED_PRESET_LABELS) as CreatedPreset[]).map((preset) => (
-            <button
-              key={preset}
-              type="button"
-              onClick={() => setCreatedPreset(preset)}
-              className={cn(
-                "w-full flex items-center justify-between gap-2 px-2 py-1.5 rounded text-sm",
-                "hover:bg-bg-surface",
-                createdPreset === preset && "text-fg font-medium",
-              )}
-            >
-              {CREATED_PRESET_LABELS[preset]}
-              {createdPreset === preset && <CheckIcon className="size-3.5 text-brand" />}
-            </button>
-          ))}
-          {createdPreset === "custom" && (
-            <div className="mt-1 pt-2 border-t border-border space-y-2 px-1 pb-1">
-              <label className="block">
-                <span className="text-xs text-fg-muted mb-1 block">From</span>
-                <Input
-                  type="date"
-                  value={msToDateInput(customAfter)}
-                  onChange={(e) => {
-                    const v = e.target.value;
-                    setCustomAfter(v ? new Date(v).getTime() : undefined);
-                  }}
-                />
-              </label>
-              <label className="block">
-                <span className="text-xs text-fg-muted mb-1 block">To</span>
-                <Input
-                  type="date"
-                  value={msToDateInput(customBefore)}
-                  onChange={(e) => {
-                    const v = e.target.value;
-                    // End of selected day so "<= that day" works as the
-                    // user expects (otherwise picking 2026-05-24 would
-                    // exclude that whole day since the filter is `<`).
-                    if (!v) {
-                      setCustomBefore(undefined);
-                      return;
-                    }
-                    const d = new Date(v);
-                    d.setDate(d.getDate() + 1);
-                    setCustomBefore(d.getTime());
-                  }}
-                />
-              </label>
-            </div>
-          )}
-        </PopoverContent>
-      </FilterChip>
+      <CreatedFilterChip value={created} onChange={setCreated} />
     </>
   );
