+++ apps/console/src/pages/Dashboard.tsx
@@ -4,8 +4,8 @@ import { useAuth } from "../lib/auth";
 import { useApiQuery } from "../lib/useApiQuery";
 import { toast } from "sonner";
 import { StatusPill } from "../components/Badge";
-import { BrandLoader } from "../components/BrandLoader";
 import { EmptyState } from "../components/EmptyState";
+import { Skeleton } from "../components/Skeleton";
 
 interface Stats {
   agents: number;
@@ -42,10 +42,6 @@ export function Dashboard() {
   );
   const stats = statsQuery.data ?? null;
   const recentSessions = sessionsQuery.data?.data.slice(0, 5) ?? [];
-  // Block initial render until BOTH first fetches settle (succeed or fail)
-  // so the page doesn't shift layout twice. `isLoading` is true only on
-  // the very first fetch; refetches stay invisible.
-  const loading = statsQuery.isLoading || sessionsQuery.isLoading;
 
   const copy = (text: string, key: string) => {
     navigator.clipboard.writeText(text);
@@ -77,14 +73,6 @@ export function Dashboard() {
     { label: "Skills", value: stats?.skills, to: "/skills" },
     { label: "Model Cards", value: stats?.model_cards, to: "/model-cards" },
   ];
-
-  if (loading) {
-    return (
-      <div className="flex-1 flex items-center justify-center">
-        <BrandLoader size="lg" label="Loading dashboard" />
-      </div>
-    );
-  }
 
   const cmd = "npx -y -p @openma/cli oma";
   const cmdGlobal = "npm i -g @openma/cli";
@@ -214,7 +202,18 @@ export function Dashboard() {
             </button>
           </div>
 
-          {recentSessions.length === 0 ? (
+          {sessionsQuery.isLoading ? (
+            <div className="border border-border rounded-lg divide-y divide-border">
+              {Array.from({ length: 4 }).map((_, i) => (
+                <div key={i} className="flex items-center gap-4 px-4 py-3">
+                  <Skeleton className="h-3.5 w-[40%]" rounded="sm" />
+                  <Skeleton className="h-3.5 w-16" rounded="sm" />
+                  <Skeleton className="h-3.5 w-32" rounded="sm" />
+                  <Skeleton className="h-3.5 w-20 ml-auto" rounded="sm" />
+                </div>
+              ))}
+            </div>
+          ) : recentSessions.length === 0 ? (
             <EmptyState
               title="No sessions yet — the stable's empty."
               body={
